Drop toBePresent, toBeEmpty. Add toExist
