--- packages/jasmine-enzyme/spec/fixtures/toBePresent.fixture.js
@@ -1,16 +1,0 @@
-'use strict';
-
-Object.defineProperty(exports, '__esModule', {
-  value: true,
-});
-exports.default = Fixture;
-var React = require('react');
-
-function Fixture() {
-  return React.createElement(
-    'div',
-    null,
-    React.createElement('span', { className: 'matches' })
-  );
-}
-module.exports = exports['default'];
